fix real entropy calc

+++ mfkdf2/src/definitions/entropy.rs
@@ -1,6 +1,6 @@
 #[cfg_attr(feature = "bindings", derive(uniffi::Record))]
-#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize, Eq, PartialEq)]
+#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize, PartialEq)]
 pub struct MFKDF2Entropy {
-  pub real:        u32,
+  pub real:        f64,
   pub theoretical: u32,
 }
